requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
requests>=2.31.0
beautifulsoup4>=4.12.0
apscheduler>=3.10.0
pyyaml>=6.0
pydantic>=2.0.0
